* 'hotfix/fixbug' of github.com:zhiyicx/thinksns-plus-android:
  (#1324) code 社区签到排行榜字段修改
  (#1324) code 问答列表样式微调
  (#1324) code 资讯评论数据调整
  (#1324) code 购买音乐更新播放队列

+++ app/src/main/java/com/zhiyicx/thinksnsplus/data/beans/DynamicListAdvert.java
@@ -160,7 +160,7 @@ public class DynamicListAdvert implements Serializable, Parcelable {
         infoListDataBean.setCategory(category);
         infoListDataBean.setUser_id(-1L);// 广告位标识
         infoListDataBean.setMaxId(max_id);// max_id 转移
-        infoListDataBean.setId((int)max_id);
+        infoListDataBean.setId(max_id);
         infoListDataBean.setTitle(advert.getTitle());// 广告内容
         infoListDataBean.setCreated_at(advert.getTime());// 广告时间
         infoListDataBean.setUpdated_at(advert.getLink());// 广告外链
